tools/hwci: move test helpers infrastructure to utils module
