--- tools/hwci/tests/__init__.py
@@ -1,10 +1,0 @@
-# Licensed under the Apache License, Version 2.0 or the MIT License.
-# SPDX-License-Identifier: Apache-2.0 OR MIT
-# Copyright Tock Contributors 2024.
-
-from .console_hello_test import (
-    OneshotTest,
-    AnalyzeConsoleTest,
-    WaitForConsoleMessageTest,
-)
-from .c_hello import test as c_hello_test
