+++ tools/hwci/utils/test_helpers/__init__.py
@@ -1,0 +1,3 @@
+from .oneshot import OneshotTest
+from .analyze_console import AnalyzeConsoleTest
+from .wait_for_console_message import WaitForConsoleMessageTest
